// typechain-types/CredentialManager.ts
export interface CredentialManagerInterface extends Interface {
  getFunction(
    nameOrSignature:
      | "getCredentialStatus"
      | "registerCredential"
      | "revokeCredential"
      | "verifyCredential"
  ): FunctionFragment;

  getEvent(
    nameOrSignatureOrTopic: "CredentialRegistered" | "CredentialRevoked"
  ): EventFragment;

  encodeFunctionData(
    functionFragment: "getCredentialStatus",
    values: [BytesLike]
  ): string;
  encodeFunctionData(
    functionFragment: "registerCredential",
    values: [BytesLike]
  ): string;
  encodeFunctionData(
    functionFragment: "revokeCredential",
    values: [BytesLike]
  ): string;
  encodeFunctionData(
    functionFragment: "verifyCredential",
    values: [BytesLike]
  ): string;

  decodeFunctionResult(
    functionFragment: "getCredentialStatus",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "registerCredential",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "revokeCredential",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "verifyCredential",
    data: BytesLike
  ): Result;
}

export namespace CredentialRegisteredEvent {
  export type InputTuple = [
    credentialHash: BytesLike,
    issuer: AddressLike,
    issuedAt: BigNumberish
  ];
  export type OutputTuple = [
    credentialHash: string,
    issuer: string,
    issuedAt: bigint
  ];
  export interface OutputObject {
    credentialHash: string;
    issuer: string;
    issuedAt: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace CredentialRevokedEvent {
  export type InputTuple = [
    credentialHash: BytesLike,
    issuer: AddressLike,
    revokedAt: BigNumberish
  ];
  export type OutputTuple = [
    credentialHash: string,
    issuer: string,
    revokedAt: bigint
  ];
  export interface OutputObject {
    credentialHash: string;
    issuer: string;
    revokedAt: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface CredentialManager extends BaseContract {
  connect(runner?: ContractRunner | null): CredentialManager;
  waitForDeployment(): Promise<this>;

  interface: CredentialManagerInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  getCredentialStatus: TypedContractMethod<
    [credentialHash: BytesLike],
    [
      [string, boolean, bigint, bigint, boolean] & {
        issuer: string;
        active: boolean;
        issuedAt: bigint;
        revokedAt: bigint;
        exists: boolean;
      }
    ],
    "view"
  >;

  registerCredential: TypedContractMethod<
    [credentialHash: BytesLike],
    [void],
    "nonpayable"
  >;

  revokeCredential: TypedContractMethod<
    [credentialHash: BytesLike],
    [void],
    "nonpayable"
  >;

  verifyCredential: TypedContractMethod<
    [credentialHash: BytesLike],
    [boolean],
    "view"
  >;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "getCredentialStatus"
  ): TypedContractMethod<
    [credentialHash: BytesLike],
    [
      [string, boolean, bigint, bigint, boolean] & {
        issuer: string;
        active: boolean;
        issuedAt: bigint;
        revokedAt: bigint;
        exists: boolean;
      }
    ],
    "view"
  >;
  getFunction(
    nameOrSignature: "registerCredential"
  ): TypedContractMethod<[credentialHash: BytesLike], [void], "nonpayable">;
  getFunction(
    nameOrSignature: "revokeCredential"
  ): TypedContractMethod<[credentialHash: BytesLike], [void], "nonpayable">;
  getFunction(
    nameOrSignature: "verifyCredential"
  ): TypedContractMethod<[credentialHash: BytesLike], [boolean], "view">;

  getEvent(
    key: "CredentialRegistered"
  ): TypedContractEvent<
    CredentialRegisteredEvent.InputTuple,
    CredentialRegisteredEvent.OutputTuple,
    CredentialRegisteredEvent.OutputObject
  >;
  getEvent(
    key: "CredentialRevoked"
  ): TypedContractEvent<
    CredentialRevokedEvent.InputTuple,
    CredentialRevokedEvent.OutputTuple,
    CredentialRevokedEvent.OutputObject
  >;

  filters: {
    "CredentialRegistered(bytes32,address,uint256)": TypedContractEvent<
      CredentialRegisteredEvent.InputTuple,
      CredentialRegisteredEvent.OutputTuple,
      CredentialRegisteredEvent.OutputObject
    >;
    CredentialRegistered: TypedContractEvent<
      CredentialRegisteredEvent.InputTuple,
      CredentialRegisteredEvent.OutputTuple,
      CredentialRegisteredEvent.OutputObject
    >;

    "CredentialRevoked(bytes32,address,uint256)": TypedContractEvent<
      CredentialRevokedEvent.InputTuple,
      CredentialRevokedEvent.OutputTuple,
      CredentialRevokedEvent.OutputObject
    >;
    CredentialRevoked: TypedContractEvent<
      CredentialRevokedEvent.InputTuple,
      CredentialRevokedEvent.OutputTuple,
      CredentialRevokedEvent.OutputObject
    >;
  };
}